Make timestamp plugin work with changesets
Changesets pass themselves as an argument to command pipelines, rather
than the raw tuple.  Because the timestamp plugin was using Hash/Array
to determine the method to use for blending new data, it was triggering
a failure when the changeset fell through the case statement here.

We should probably keep an eye out for any similar patterns in command
hooks as the resulting error message was an utterly unclear:
      ArgumentError:
       wrong number of arguments (given 0, expected 1)
     # /home/flip/.gem/ruby/2.4.2/gems/rom-sql-2.3.0/lib/rom/sql/commands/create.rb:22:in `execute'