rename so as not to clash with the existing method in rom-sql
From the looks of it, there are parts of rom-sql that have already
internalized the idea that with_input_tuples will _always_ return an
array.  Fortunately, those look like they're at the _end_ of the input
chain -- multi_insert and such -- and not part of the user hook cycle,
which was the impetus for this branch in the first place.